Fixed Duplicate Field, and Framework Files
The data file had a duplicate field and needed to be fixed.  It was
necessary to push the fixed data file to Filebin.ca, therefore the
'inputDatasetURL environment variable needed to be updated with new URL
in the data prep file, line 30.  Also lines 94, and 203 needed to be
updated to the new field name.

Finally, I added a semicolon to the framework individual data analysis
file, as it needed a final semicolon.